requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.20.1
autopep8==1.6.0
backcall==0.2.0
bottle==0.12.21
bottle-websocket==0.2.9
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
colour==0.1.5
commonmark==0.9.1
comtypes==1.1.11
cycler==0.11.0
decorator==5.1.0
Eel==0.14.0
fonttools==4.28.5
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
glcontext==2.3.4
googletrans==4.0.0rc1
greenlet==1.1.2
gTTS==2.2.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==7.30.0
jedi==0.18.1
kiwisolver==1.3.2
manimgl==1.3.0
ManimPango==0.3.1
mapbox-earcut==1.0.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
moderngl==5.6.4
moderngl-window==2.4.1
mpmath==1.2.1
multipledispatch==0.6.0
numpy==1.21.4
opencv-contrib-python==4.5.5.62
opencv-python==4.5.4.60
packaging==21.3
parso==0.8.2
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.0
playsound==1.2.2
prompt-toolkit==3.0.23
pycodestyle==2.8.0
pycparser==2.21
pydub==0.25.1
pyglet==1.5.21
Pygments==2.10.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
PyOpenGL==3.1.5
pyparsing==3.0.6
PyPDF2==2.0.0
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyreadline==2.1
pyrr==0.10.3
pytesseract==0.3.9
python-dateutil==2.8.2
pyttsx3==2.90
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rich==11.0.0
scipy==1.7.3
screeninfo==0.8
six==1.16.0
skia-pathops==0.7.2
sniffio==1.2.0
sympy==1.9
tk==0.1.0
toml==0.10.2
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.9
validators==0.18.2
wcwidth==0.2.5
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
